+++ src/element.rs
@@ -21,7 +21,7 @@ use crate::traits::{Cloneable, None, Trait};
 /// # Consuming
 ///
 /// Whenever you have `ElementPointer` as a value (from destructive [`AnyVec`] operations),
-/// you can safely take pointed value, with [`AnyValue::downcast`] or [`any_value::move_into`].
+/// you can safely take pointed value, with [`AnyValue::downcast`] or [`any_value::move_out`].
 /// Otherwise, it will be destructed with destruction of `Element`.
 ///
 /// # Notes
@@ -34,7 +34,7 @@ use crate::traits::{Cloneable, None, Trait};
 /// [`drain`]: crate::AnyVec::drain
 /// [`splice`]: crate::AnyVec::splice
 /// [`any_value`]: crate::any_value
-/// [`any_value::move_into`]: crate::any_value::move_into
+/// [`any_value::move_out`]: crate::any_value::move_out
 pub struct ElementPointer<'a, AnyVecPtr: IAnyVecRawPtr>{
     any_vec_ptr: AnyVecPtr,
     element: NonNull<u8>,